movie_app_backend/requirements.txt
Django == 4.2.1
djangorestframework== 3.14.0
django-cors-headers == 4.2.0
#mysqlclient ==  2.2.0